refactor!: rename `eth` to `ethernet`
