--- src/ariel-os-embassy/src/eth.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "ethernet-stm32")]
-pub(crate) use crate::hal::eth::NetworkDevice;
