+++ src/ariel-os-embassy/src/ethernet.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "ethernet-stm32")]
+pub(crate) use crate::hal::ethernet::NetworkDevice;
