+++ src/ariel-os-stm32/Cargo.toml
@@ -78,8 +78,8 @@ storage = ["dep:embassy-embedded-hal"]
 ## Enables USB support.
 usb = []
 
-eth = []
-ethernet-stm32 = ["dep:embassy-embedded-hal", "eth"]
+ethernet = []
+ethernet-stm32 = ["dep:embassy-embedded-hal", "ethernet"]
 
 ## Enables defmt support.
 defmt = ["dep:defmt", "embassy-stm32/defmt"]
