+++ src/ariel-os-stm32/src/ethernet.rs
